Palette swapping and classes ImageData and BiteImage merged.
Several palettes can be added to sprite sheets and be used by their sprites.

BiteImage took over ImageData's pixel map storage. Separating the two turned out to be pointless and limiting.

+++ src/framebuffer.cpp
@@ -71,8 +71,8 @@ namespace Bite
 			glDrawBuffers( 1, drawbuffs );
 			SetForRead();
 
-			int dstWidth =  ResWidth() * PixSize();
-			int dstHeight = ResHeight() * PixSize();
+			int dstWidth = int( ResWidth() * PixSize() );
+			int dstHeight = int( ResHeight() * PixSize() );
 			int dstLeft = WindowWidth()/2 - dstWidth/2;
 			int dstBottom = WindowHeight()/2 - dstHeight/2;
 			int dstRight = dstLeft + dstWidth;
